+++ kinto_http/__init__.py
@@ -8,13 +8,15 @@ import logging
 from kinto_http import utils
 from kinto_http.session import create_session, Session
 from kinto_http.batch import BatchSession
-from kinto_http.exceptions import BucketNotFound, CollectionNotFound, KintoException
+from kinto_http.exceptions import (
+    BucketNotFound, CollectionNotFound, KintoException, KintoBatchException
+)
 from kinto_http.patch_type import PatchType, BasicPatch
 
 logger = logging.getLogger('kinto_http')
 
 __all__ = ('Endpoints', 'Session', 'Client', 'create_session',
-           'BucketNotFound', 'CollectionNotFound', 'KintoException')
+           'BucketNotFound', 'CollectionNotFound', 'KintoException', 'KintoBatchException')
 
 
 OBJECTS_PERMISSIONS = {