+++ CHANGELOG.rst
@@ -7,6 +7,12 @@ This document describes changes between each past release.
 9.2.0 (unreleased)
 ==================
 
+**Breaking changes**
+
+By default, the client now raises an exception when a 4XX error occurs in a batch request (#154)
+
+In order to ignore those errors as before, instantiate the client with ``ignore_batch_4xx=True``.
+
 **New feature**
 
 - Raise a specific `CollectionNotFound` exception rather than a generic `KintoException`.